Bugfix: remove yellow artifact on Windows server
git-svn-id: file:///home/code/svn/ovd/trunk@5101 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ client/java/src/net/propero/rdp/Bitmap.java
@@ -1305,7 +1305,7 @@ public class Bitmap {
 								}
 								if ((mask & mixmask) != 0)
 									// setli(pixel, line, x, mix, Bpp);
-									pixel[line + x] = (byte) mix;
+									pixel[line + x] = mix;
 								else
 									// setli(pixel, line, x, 0, Bpp);
 									pixel[line + x] = 0;
